Create Guess the number Game
This is a game where the algorithm generates a random number and you cannot see it. You guess it. With each guess your guess count increases. If you have entered a bigger or smaller number than expected it will tell you. The game will end when you guess the number correctly.